refactor: introduce PathResolver
